refactor: adds file comments and groups shared components
