+++ src/components/Footer/index.tsx
@@ -2,6 +2,7 @@
 
 import { FooterStyled } from '@/components/Footer/styles';
 
+// The component to display a page footer.
 const Footer = () => (
   <FooterStyled>
     <p>Facto Expense 2025 - All rights reserved</p>
