+++ src/components/LoginForm/index.tsx
@@ -1,11 +1,13 @@
 'use client'
 
-import Button from '@/components/Button';
-import FieldText from '@/components/FieldText';
-import { useLogin } from '@/domain/user/client';
+import Button from '@/components/shared/Button';
+import FieldText from '@/components/shared/FieldText';
+import { useLogin } from '@/domain/user/api/client';
 import { Credentials } from '@/domain/user/types';
 import { Form, Formik } from 'formik';
 
+// The component to display the login form.
+// Log in is done by sending username and password values through the API.
 const LoginForm = () => {
   const { mutate: login, isError } = useLogin();
 
@@ -40,6 +42,7 @@ const LoginForm = () => {
               error={showError}
               onChange={handleChange}
             />
+            {showError && <><p>Invalid credentials</p><br /></>}
             <Button type="submit">Log in</Button>
           </Form>
         )
