+++ src/components/TenantSelect/index.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { setTenantId } from '@/app/actions';
-import FieldSelect from '@/components/FieldSelect';
+import FieldSelect from '@/components/shared/FieldSelect';
 import { tenants } from '@/core/constants';
 import { ChangeEvent, useEffect } from 'react';
 
@@ -9,6 +9,7 @@ interface HeaderProps {
   initialValue?: string;
 }
 
+// The component to display the tenant selector.
 const TenantSelect = ({ initialValue }: HeaderProps) => {
   const setTenant = (evt: ChangeEvent<HTMLSelectElement>) => {
     const tenantId = evt.target.value;
