+++ src/components/shared/Table/styles.ts
